Split out ValveMotorDirectV1HardwareDriverBase
