+++ content/OTRadioLink/OTRadValve.h
@@ -33,6 +33,8 @@ Author(s) / Copyright (s): Damon Hart-Davis 2015--2016
 // Driver for FHT8V wireless valve actuator (and FS20 protocol encode/decode).
 #include "utility/OTRadValve_FHT8VRadValve.h"
 
+#include "utility/OTRadValve_ValveMotorBase.h"
+
 // Driver for DORM1/REV7 direct motor drive.
 #include "utility/OTRadValve_ValveMotorDirectV1.h"
