+++ content/OTRadioLink/utility/OTRadValve_ValveMotorBase.cpp
@@ -1,0 +1,92 @@
+/*
+ * OTRadValve_ValveMotorBase.cpp
+ *
+ *  Created on: 11 Oct 2016
+ *      Author: denzo
+ */
+
+
+#include "OTRadValve_AbstractRadValve.h"
+#include "OTRadValve_ValveMotorBase.h"
+
+//#include "OTV0P2BASE_Serial_IO.h"
+
+namespace OTRadValve
+{
+
+// Spin for up to the specified number of SCT ticks, monitoring current and position encoding.
+//   * maxRunTicks  maximum sub-cycle ticks to attempt to run/spin for); strictly positive
+//   * minTicksBeforeAbort  minimum ticks before abort for end-stop / high-current,
+//       don't attempt to run at all if less than this time available before (close to) end of sub-cycle;
+//       should be no greater than maxRunTicks
+//   * dir  direction to run motor (open or closed) or off if waiting for motor to stop
+//   * callback  handler to deliver end-stop and position-encoder callbacks to;
+//     non-null and callbacks must return very quickly
+// If too few ticks remain before the end of the sub-cycle for the minimum run,
+// then this will return true immediately.
+// Invokes callbacks for high current (end stop) and position (shaft) encoder where applicable.
+// Aborts early if high current is detected at the start,
+// or after the minimum run period.
+// Returns true if aborted early from too little time to start, or by high current (assumed end-stop hit).
+bool ValveMotorDirectV1HardwareDriverBase::spinSCTTicks(const uint8_t maxRunTicks, const uint8_t minTicksBeforeAbort, const OTRadValve::HardwareMotorDriverInterface::motor_drive dir, OTRadValve::HardwareMotorDriverInterfaceCallbackHandler &callback)
+  {
+  // Sub-cycle time now.
+  const uint8_t sctStart = OTV0P2BASE::getSubCycleTime();
+  // Only run up to ~90% point of the minor cycle to leave time for other processing.
+  uint8_t sct = sctStart;
+  const uint8_t maxTicksBeforeAbsLimit = (sctAbsLimit - sct);
+  // Abort immediately if not enough time to do minimum run.
+  if((sct >= sctAbsLimit) || (maxTicksBeforeAbsLimit < minTicksBeforeAbort)) { return(true); }
+  // Note if opening or closing...
+  const bool stopped = (HardwareMotorDriverInterface::motorOff == dir);
+  const bool isOpening = (HardwareMotorDriverInterface::motorDriveOpening == dir);
+  bool currentHigh = false;
+  // Compute time minimum time before return, then target time before stop/return.
+  const uint8_t sctMinRunTime = sctStart + minTicksBeforeAbort; // Min run time to avoid false readings.
+  const uint8_t sctMaxRunTime = sctStart + min(maxRunTicks, maxTicksBeforeAbsLimit);
+  // Do minimum run time, NOT checking for end-stop / high current.
+  for( ; ; )
+    {
+    // Poll shaft encoder output and update tick counter.
+    const uint8_t newSct = OTV0P2BASE::getSubCycleTime();
+    if(newSct != sct)
+      {
+      sct = newSct; // Assumes no intermediate values missed.
+      if(!stopped) { callback.signalRunSCTTick(isOpening); }
+      if(sct >= sctMinRunTime) { break; }
+      }
+    // TODO: shaft encoder
+    }
+
+  // Do as much of requested above-minimum run-time as possible,
+  // iff run time beyond the minimum was actually requested
+  // (else avoid the current sampling entirely).
+  if(sctMaxRunTime > sctMinRunTime)
+    {
+    for( ; ; )
+      {
+      // Check for high current and abort if detected.
+      if(isCurrentHigh(dir)) { currentHigh = true; break; }
+      // Poll shaft encoder output and update tick counter.
+      const uint8_t newSct = OTV0P2BASE::getSubCycleTime();
+      if(newSct != sct)
+        {
+        sct = newSct; // Assumes no intermediate values missed.
+        if(!stopped) { callback.signalRunSCTTick(isOpening); }
+        if(sct >= sctMaxRunTime) { break; }
+        }
+      }
+    }
+
+  // Call back and return true if current high / end-stop seen.
+  if(currentHigh)
+    {
+    callback.signalHittingEndStop(isOpening);
+    return(true);
+    }
+  return(false);
+  }
+
+
+}
+
